+++ packages/renderer/threejs-background/src/background-manager/index.ts
@@ -1,3 +1,1 @@
-export * from "./star-field";
 export * from "./debug-visualizer";
-export * from "./parallax-handler";