+++ packages/renderer/threejs-background/src/index.ts
@@ -2,4 +2,11 @@ export * from "./BackgroundManager";
 
 export * from "./background-manager";
 
+export * from "./fields/core/Field";
+export * from "./fields/star-field/StarField";
+export * from "./fields/star-field/types";
+export * from "./fields/nebula-field/NebulaField";
+export * from "./fields/nebula-field/types";
+export * from "./fields/nebula-field/palettes";
+
 export {};