+++ packages/renderer/threejs-background/vite.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig } from "vite";
+
+export default defineConfig({
+  plugins: [
+    {
+      name: "glsl",
+      transform(src, id) {
+        if (/\.glsl$/.test(id)) {
+          return {
+            code: `export default ${JSON.stringify(src)};`,
+            map: null,
+          };
+        }
+      },
+    },
+  ],
+});